Fix proptypes add isRequred Header
